fix: extract actix dependency from near-primitives
