+++ chain/client/Cargo.toml
@@ -34,6 +34,8 @@ near-chunks = { path = "../chunks" }
 near-telemetry = { path = "../telemetry" }
 
 [dev-dependencies]
+near-env = { path = "../../env" }
+testlib = { path = "../../test-utils/testlib" }
 neard = { path = "../../neard" }
 
 [features]
