+++ chain/client/tests/bug_repros.rs
@@ -10,10 +10,10 @@ use near_chain::test_utils::account_id_to_shard_id;
 use near_client::test_utils::setup_mock_all_validators;
 use near_client::{ClientActor, ViewClientActor};
 use near_crypto::{InMemorySigner, KeyType};
+use near_env::init_test_logger;
 use near_network::types::NetworkRequests::PartialEncodedChunkMessage;
 use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses, PeerInfo};
 use near_primitives::block::Block;
-use near_primitives::test_utils::init_test_logger;
 use near_primitives::transaction::SignedTransaction;
 use std::cmp::max;
 use std::collections::HashMap;
