+++ chain/client/tests/cross_shard_tx.rs
@@ -6,8 +6,8 @@ use futures::{future, FutureExt};
 
 use near_client::test_utils::setup_mock_all_validators;
 use near_client::{ClientActor, Query, ViewClientActor};
+use near_env::init_test_logger;
 use near_network::{NetworkRequests, NetworkResponses, PeerInfo};
-use near_primitives::test_utils::init_test_logger;
 use near_primitives::types::BlockIdOrFinality;
 use near_primitives::views::{QueryRequest, QueryResponseKind::ViewAccount};
 
@@ -87,11 +87,11 @@ mod tests {
     use near_client::test_utils::setup_mock_all_validators;
     use near_client::{ClientActor, Query, ViewClientActor};
     use near_crypto::{InMemorySigner, KeyType};
+    use near_env::init_test_logger;
     use near_network::{
         NetworkClientMessages, NetworkClientResponses, NetworkRequests, NetworkResponses, PeerInfo,
     };
     use near_primitives::hash::CryptoHash;
-    use near_primitives::test_utils::init_test_logger;
     use near_primitives::transaction::SignedTransaction;
     use near_primitives::types::{AccountId, BlockIdOrFinality};
     use near_primitives::views::{QueryRequest, QueryResponse, QueryResponseKind::ViewAccount};
