+++ chain/jsonrpc/tests/rpc_query.rs
@@ -5,13 +5,13 @@ use futures::{future, FutureExt};
 
 use near_chain_configs::PROTOCOL_VERSION;
 use near_crypto::{KeyType, PublicKey, Signature};
+use near_env::init_test_logger;
 use near_jsonrpc::client::new_client;
 use near_jsonrpc_client::ChunkId;
 use near_network::test_utils::WaitOrTimeout;
 use near_primitives::account::{AccessKey, AccessKeyPermission};
 use near_primitives::hash::CryptoHash;
 use near_primitives::rpc::{RpcGenesisRecordsRequest, RpcPagination, RpcQueryRequest};
-use near_primitives::test_utils::init_test_logger;
 use near_primitives::types::{BlockId, BlockIdOrFinality, Finality, ShardId};
 use near_primitives::views::{QueryRequest, QueryResponseKind};
