+++ chain/jsonrpc/tests/rpc_transactions.rs
@@ -5,11 +5,11 @@ use borsh::BorshSerialize;
 use futures::{future, FutureExt, TryFutureExt};
 
 use near_crypto::{InMemorySigner, KeyType};
+use near_env::{init_integration_logger, init_test_logger};
 use near_jsonrpc::client::new_client;
 use near_network::test_utils::WaitOrTimeout;
 use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::serialize::{to_base, to_base64};
-use near_primitives::test_utils::{init_integration_logger, init_test_logger};
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::BlockIdOrFinality;
 use near_primitives::views::FinalExecutionStatus;
