+++ chain/network/Cargo.toml
@@ -31,6 +31,7 @@ near-metrics = { path = "../../core/metrics" }
 near-chain = { path = "../chain" }
 
 [dev-dependencies]
+near-env = {path = "../../env"}
 tempdir = "0.3.7"
 bencher = "0.1.5"
 num-rational = "0.2.4"
