+++ neard/tests/sync_state_nodes.rs
@@ -7,9 +7,10 @@ use tempdir::TempDir;
 
 use near_chain_configs::Genesis;
 use near_client::GetBlock;
+use near_env::init_integration_logger;
 use near_network::test_utils::{convert_boot_nodes, open_port, WaitOrTimeout};
-use near_primitives::test_utils::{heavy_test, init_integration_logger};
 use neard::{config::GenesisExt, load_test_config, start_with_config};
+use testlib::test_helpers::heavy_test;
 
 /// One client is in front, another must sync to it using state (fast) sync.
 #[test]
