+++ test-utils/state-viewer/src/main.rs
@@ -6,12 +6,12 @@ use ansi_term::Color::Red;
 use clap::{App, Arg, SubCommand};
 
 use near_chain::{ChainStore, ChainStoreAccess, RuntimeAdapter};
+use near_env::init_integration_logger;
 use near_network::peer_store::PeerStore;
 use near_primitives::block::BlockHeader;
 use near_primitives::hash::CryptoHash;
 use near_primitives::serialize::to_base;
 use near_primitives::state_record::StateRecord;
-use near_primitives::test_utils::init_integration_logger;
 use near_primitives::types::{BlockHeight, StateRoot};
 use near_store::test_utils::create_test_store;
 use near_store::{create_store, Store, TrieIterator};
