+++ test-utils/testlib/Cargo.toml
@@ -23,6 +23,7 @@ num-rational = "0.2.4"
 
 borsh = "0.6.1"
 
+near-env = { path = "../../env" }
 near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
